+++ backend/src/lib.rs
@@ -2,6 +2,6 @@ pub mod app;
 pub mod auth;
 pub mod controllers;
 pub mod entities;
-pub mod mongo;
+pub mod repositories;
 
 pub use app::{App, Args, Error};
