+++ backend/src/repositories/mod.rs
@@ -1,0 +1,5 @@
+mod mongo;
+mod traits;
+
+pub use mongo::*;
+pub use traits::*;
